requirements.txt
tqdm==4.64.0
openslide-python==1.2.0
numpy==1.22.4
matplotlib==3.5.2
torch==1.11.0+cpu
torchaudio==0.11.0+cpu
torchvision==0.12.0+cpu
jupyterlab==3.4.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
pyFAST==4.4.0
pyvips==2.2.1
Pillow==9.1.1
segmentation-models-pytorch==0.2.1
